Refactor: dto 및 entity 구분
