+++ src/main/java/com/example/oliveyoungbe/entity/TicketBooking.java
@@ -1,8 +1,9 @@
-package com.example.oliveyoungbe.dto;
+package com.example.oliveyoungbe.entity;
 
 import jakarta.annotation.Nullable;
 import lombok.Data;
 
+import javax.persistence.Entity;
 import java.sql.Timestamp;
 
 @Entity
